add enableVUART module parameter
This patch adds a new parameter to the kernel module that allows user-space to
control whether the driver handles packets sent to the VUART I/O ranges.
If enableVUART is set to 1 (= on), all core I/O to the VUART ranges is handled
by mapping it to the driver-maintained virtual devices. User-space applications
will not receive such packets via the usual interface.
If enableVUART is set to 0 (= off), the driver ignores packets to the VUART
ranges and delivers them to the application.